fix(auth): refactor authentication context and provider, update imports in main.tsx
