+++ FRONTEND/src/context/AuthContext.ts
@@ -1,0 +1,13 @@
+import { createContext } from "react";
+
+interface AuthContextType {
+    accessToken: string | null;
+    setAccessToken: (token: string | null) => void;
+    logout: () => void;
+}
+
+export const AuthContext = createContext<AuthContextType>({
+    accessToken: null,
+    setAccessToken: () => {},
+    logout: () => {},
+});
