+++ FRONTEND/src/context/AuthProvider.tsx
@@ -1,21 +1,6 @@
-import { createContext, useContext, useState, type ReactNode } from "react";
 import { setAccessToken as storeSetAccessToken } from '../auth/tokenStore'
-
-interface AuthContextType {
-    accessToken: string | null;
-    setAccessToken: (token: string | null) => void;
-    logout: () => void;
-}
-
-const AuthContext = createContext<AuthContextType>({
-    accessToken: null,
-    setAccessToken: () => {},
-    logout: () => {},
-});
-
-export const useAuth = () => {
-    return useContext(AuthContext);
-};
+import { useState, type ReactNode } from "react";
+import { AuthContext } from "./AuthContext";
 
 export const AuthProvider = ({ children }: { children: ReactNode }) => {
     const [accessToken, setAccessToken] = useState<string | null>(null);
