+++ FRONTEND/src/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import { AuthProvider } from './hooks/useAuth.tsx'
+import { AuthProvider } from './context/AuthProvider.tsx'
 import './index.css'
 import App from './App.tsx'
